OVSDriver: add experimental packet pipeline
This code, enabled with the --pipeline option, hardcodes the T6 forwarding
model. All L2 logic is implemented, minus LAGs and multicast.

OFTest test suite included.

Most of this code will never be merged into the opensource repo because it
hardcodes details of our product and the Broadcom chip. The plan is to get this
fixed pipeline fully working, then make it programmable, and then merge it to
opensource.