pipeline_bvs: support bsn_vlan_xlate_port_group_id
The port table is changed to store the new field in its value. The vlan_xlate
and egr_vlan_xlate tables are changed to accept the new field in the key. A
type field is added to the vlan xlate keys to differentiate between entries
keyed on port group and entries keyed on lag id or port number. Finally, the
main pipeline code is changed to look up using the port group first before
falling back to the old lag id or port number keys.

All of the backwards compatibility code will be removed once we've finished
porting OFTest and given the controller team time to update.