nat: network namespace lifecycle
When a NAT table entry is added, a new network namespace is created. IVS can
use the file descriptor returned by create_netns() to re-enter the namespace.
When a NAT table entry is deleted, IVS calls close() on the netns fd, which
drops the last reference count on the network namespace in the kernel. The
kernel will clean up any veths with one half in the namespace, as well as
cleaning up all the routing table entries, iptables rules, etc.