nat: implement NAT
This is a port of the NAT prototype [1] to IVS. Most of the work is done by the
"ip" and "iptables" commands that we shell out to. This makes implementation
much easier, but it's also a little slow (50 ms to insert a NAT table entry).
In the future we may implement more of these functions in C for better
performance and reduced reliance on the host OS.

See the IVS NAT wiki page [2] for documentation on the NAT scheme.

[1] https://github.com/bigswitch/nat-prototype
[2] https://bigswitch.atlassian.net/wiki/display/XE/PAT