pipeline_bvs: add counters for important codepaths
These counters are exposed in a new generic stats message. I added counters to
every place we drop the packet and a few other important paths.

These aren't using the debug counter infrastructure because they need to count
packets that hit in the kernel flowtable, not just ones that go to userspace.

Future work:
 - More efficient encoding
 - Per-port stats