nat: improve reliability of modify (BVS-3614)
When just the IP was changed (leaving the interface names the same) IVS didn't
receive vport-deleted notifications from the kernel in a timely fashion, and it
never received vport-add notifications for the new ports. Adding explicit calls
to indigo_port_interface_remove gets us the vport-deleted notifications
synchronously instead of waiting for the network namespace to be torn down.