host_stats: fix meminfo parsing for new kernel versions
Linux 3.14 introduced the MemAvailable line. This broke our parsing of the
meminfo file since we were using a fixed scanf string.

The new code accepts the fields in any order and shouldn't break when new
fields are added.

The new MemAvailable field is a better estimate of free memory than
free+cached, so we use it if it's there.